move integration tests into separate directory
