+++ tests/core/case/mod.rs
@@ -1,0 +1,1 @@
+mod string;
