+++ tests/core/mod.rs
@@ -1,0 +1,2 @@
+mod case;
+mod file;
