+++ tests/tests.rs
@@ -1,0 +1,3 @@
+mod arguments;
+mod commands;
+mod core;
